+++ mayan/apps/dependencies/literals.py
@@ -1,0 +1,6 @@
+from __future__ import print_function, unicode_literals
+
+DEFAULT_LOCK_FILENAME = 'package-lock.json'
+DEFAULT_PACKAGE_FILENAME = 'package.json'
+DEFAULT_MODULE_DIRECTORY = 'node_modules'
+DEFAULT_REGISTRY_URL = 'http://registry.npmjs.com'